+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Fixed the mod menu sometimes being blank when reopening it
 - Block exiting from the mod install menu with keys such as escape
+- Don't block dependency mod uninstalling when all parent mods have been uninstalled in the current session
 
 ## [0.9.18] 2024-08-08